refactor(extensions): extract the plugins and MCP modal farms
The remaining 12 GlassModals + 2 OAuth wizards (~1,000 lines of JSX)
move out of ExtensionsPanel verbatim, split by domain to stay under
the files_ge_1000 budget:

- ExtensionsPanelPluginsModals — recommended install, path install +
  validate, install confirm, plugin details + row menu, sources,
  uninstall confirm, validate result
- ExtensionsPanelMcpModals — server add form, remove confirm, doctor
  report, both OAuth wizards

extensionsSurface.guard.test.ts updated in the same PR: the surface
boundary now ends at the plugins farm component, and the
local-path-install assertions span panel + farm (the card moved with
its modal; the anchor, picker wiring and state stay on the panel).

ExtensionsPanel: 3,980 → 3,058 lines. Completes the modal-farm
extraction; only the tab surfaces remain in the panel.
