+++ src/lib/extensionsSurface.guard.test.ts
@@ -8,7 +8,7 @@ describe("extensions settings surface guard", () => {
   it("wraps every extension tab in the shared opaque settings card", () => {
     const component = read("../components/ExtensionsPanel.tsx");
     const surface = component.match(
-      /<div className="settings-card ext-panel__surface">([\s\S]*?)\n\s*<\/div>\n\n\s*<GlassModal/,
+      /<div className="settings-card ext-panel__surface">([\s\S]*?)\n\s*<\/div>\n\n\s*<ExtensionsPanelPluginsModals/,
     )?.[1];
 
     expect(surface).toBeDefined();
@@ -19,11 +19,15 @@ describe("extensions settings surface guard", () => {
 
   it("keeps local-path plugin install on the plugins Discover + control", () => {
     const component = read("../components/ExtensionsPanel.tsx");
+    const modals = read("../components/ExtensionsPanelPluginsModals.tsx");
+    // Anchor stays on the panel's Discover surface; the card + wiring moved
+    // with the plugins modal farm (WP: giant-component decomposition).
     expect(component).toContain("settings-anchor-ext-plugins-install");
-    expect(component).toContain("PluginPathInstallCard");
-    expect(component).toContain("pickDirectory");
     expect(component).toContain("openPathInstall");
     expect(component).toContain("pathInstallOpen");
+    expect(component).toContain("pickDirectory");
+    expect(modals).toContain("PluginPathInstallCard");
+    expect(modals).toContain("openPathInstall");
   });
 
   it("reuses the shared card material and keeps its modifier layout-only", () => {
